Add --filesystem=xdg-run/gvfsd:rw
I'm not sure if there are reasons not to add this, just proposing in
case it was missed accidentally.

Without this, on Fedora 36 at least, the following warning appears
when clicking a link in a mail to open in a browser:

> (evolution.bin:25): GVFS-WARNING **: 10:27:18.223: The peer-to-peer connection failed: Permission denied. Falling back to the session bus. Your application is probably missing --filesystem=xdg-run/gvfsd privileges.

It might just be me, but it seems like with this, links open slightly
snappier too.